fix wrong code : wrong url __init__.py fixed
